+++ tests/components/channels/test_digital_channel.py
@@ -1,7 +1,7 @@
 from copy import deepcopy
 from quam.components import Channel, DigitalOutputChannel, pulses
 from quam.core import QuamRoot, quam_dataclass
-from quam.core.qua_config_template import qua_config_template
+from quam.core.quam_instantiation import instantiate_quam_class
 
 
 @quam_dataclass
@@ -9,7 +9,7 @@ class QuamTest(QuamRoot):
     channel: Channel
 
 
-def test_digital_only_channel():
+def test_digital_only_channel(qua_config):
     channel = Channel(
         id="channel",
         digital_outputs={"1": DigitalOutputChannel(opx_output=("con1", 1))},
@@ -18,16 +18,15 @@ def test_digital_only_channel():
     quam = QuamTest(channel=channel)
     cfg = quam.generate_config()
 
-    expected_cfg = deepcopy(qua_config_template)
-    expected_cfg["controllers"] = {"con1": {"digital_outputs": {1: {}}}}
-    expected_cfg["elements"] = {
+    qua_config["controllers"] = {"con1": {"digital_outputs": {1: {}}}}
+    qua_config["elements"] = {
         "channel": {"digitalInputs": {"1": {"port": ("con1", 1)}}, "operations": {}}
     }
 
-    assert cfg == expected_cfg
+    assert cfg == qua_config
 
 
-def test_digital_only_pulse():
+def test_digital_only_pulse(qua_config):
     channel = Channel(
         id="channel",
         operations={
@@ -39,18 +38,24 @@ def test_digital_only_pulse():
     quam = QuamTest(channel=channel)
     cfg = quam.generate_config()
 
-    expected_cfg = deepcopy(qua_config_template)
-    # expected_cfg["controllers"] = {"con1": {"digital_outputs": {1: {}}}}
-    expected_cfg["elements"] = {
+    qua_config["elements"] = {
         "channel": {"operations": {"digital": "channel.digital.pulse"}}
     }
-    expected_cfg["pulses"]["channel.digital.pulse"] = {
+    qua_config["pulses"]["channel.digital.pulse"] = {
         "length": 100,
         "operation": "control",
         "digital_marker": "channel.digital.dm",
     }
-    expected_cfg["digital_waveforms"]["channel.digital.dm"] = {
+    qua_config["digital_waveforms"]["channel.digital.dm"] = {
         "samples": [(1, 20, 0, 10)]
     }
 
-    assert cfg == expected_cfg
+    assert cfg == qua_config
+
+
+def test_instantiate_digital_channel():
+    channel = instantiate_quam_class(
+        quam_class=DigitalOutputChannel, contents={"opx_output": ["con1", 1]}
+    )
+
+    assert channel.opx_output == ("con1", 1)