+++ tests/components/test_octave.py
@@ -4,7 +4,6 @@ import pytest
 from quam.components.channels import IQChannel, InOutIQChannel, InOutSingleChannel
 
 from quam.components.octave import Octave, OctaveUpConverter, OctaveDownConverter
-from quam.core.qua_config_template import qua_config_template
 from quam.core.quam_classes import QuamRoot, quam_dataclass
 
 
@@ -27,12 +26,11 @@ def test_instantiate_octave(octave):
     assert octave.loopbacks == []
 
 
-def test_empty_octave_config(octave):
+def test_empty_octave_config(octave, qua_config):
     machine = OctaveQuAM(octave=octave)
     config = machine.generate_config()
 
-    expected_cfg = deepcopy(qua_config_template)
-    expected_cfg["octaves"] = {
+    qua_config["octaves"] = {
         "octave1": {
             "RF_outputs": {},
             "RF_inputs": {},
@@ -41,7 +39,7 @@ def test_empty_octave_config(octave):
         }
     }
 
-    assert config == expected_cfg
+    assert config == qua_config
 
 
 def test_empty_octave_empty_config(octave):
@@ -267,7 +265,7 @@ def test_instantiate_octave_default_connectivity(octave):
         assert RF_input.id == idx
 
 
-def test_channel_add_RF_outputs(octave):
+def test_channel_add_RF_outputs(octave, qua_config):
     OctaveQuAM(octave=octave)
     octave.RF_outputs[2] = OctaveUpConverter(id=2, LO_frequency=2e9)
 
@@ -278,8 +276,7 @@ def test_channel_add_RF_outputs(octave):
         frequency_converter_up="#/octave/RF_outputs/2",
     )
 
-    cfg = deepcopy(qua_config_template)
-    channel.apply_to_config(cfg)
+    channel.apply_to_config(qua_config)
 
     expected_cfg_elements = {
         "ch": {
@@ -289,10 +286,10 @@ def test_channel_add_RF_outputs(octave):
         }
     }
 
-    assert cfg["elements"] == expected_cfg_elements
+    assert qua_config["elements"] == expected_cfg_elements
 
 
-def test_channel_add_RF_inputs(octave):
+def test_channel_add_RF_inputs(octave, qua_config):
     OctaveQuAM(octave=octave)
     octave.RF_outputs[3] = OctaveUpConverter(id=3, LO_frequency=2e9)
     octave.RF_inputs[4] = OctaveDownConverter(id=4, LO_frequency=2e9)
@@ -307,8 +304,7 @@ def test_channel_add_RF_inputs(octave):
         frequency_converter_down="#/octave/RF_inputs/4",
     )
 
-    cfg = deepcopy(qua_config_template)
-    channel.apply_to_config(cfg)
+    channel.apply_to_config(qua_config)
 
     expected_cfg_elements = {
         "ch": {
@@ -321,7 +317,7 @@ def test_channel_add_RF_inputs(octave):
         }
     }
 
-    assert cfg["elements"] == expected_cfg_elements
+    assert qua_config["elements"] == expected_cfg_elements
 
 
 def test_load_octave(octave):